add option to delete

+++ booking/src/app/components/all-trips/trip-card/trip-card.ts
@@ -3,6 +3,7 @@ import { Trip } from "../../../models/trip_model"
 import { UserService } from '../../../services/User.service';
 import { TripService } from '../../../services/Trip.service';
 import { RouterLink } from "@angular/router";
+import { BookingService } from '../../../services/Booking.service';
 
 @Component({
   selector: 'app-trip-card',
@@ -13,17 +14,26 @@ import { RouterLink } from "@angular/router";
 export class TripCard {
   private userService = inject(UserService);
   private tripService = inject(TripService);
+  private bookingService = inject(BookingService);
   currentUser = this.userService.currentUser();
   isAdmin = this.currentUser?.isAdmin;
   trip = input<Trip>();
   toDelete = signal<boolean>(false);
   deleted = output<string>();
+  errorToDelete = signal<string>('');
   
   deleteTrip(){
    const id = this.trip()?.id;
     if (!id) return;
-    this.tripService.deleteTripByID(id).subscribe(() => {
-      this.deleted.emit(id); // מעדכן את האב סיגנלית
+
+    this.bookingService.getNumberOfRegistrations(this.trip()?.id||"").subscribe((registrations) => {
+      if (registrations.length === 0) {
+        this.tripService.deleteTripByID(id).subscribe(() => {
+          this.deleted.emit(id);
+        });
+      } else {
+        this.errorToDelete.set('Cannot delete trip with existing registrations.');
+      }
     });
   }
